store JWT as a Bearer token in HTTP header
This is changed as the cookies are quite restrictive when
developing as it requires modifying your computer's `hosts`
file which is not a great experience. Also, being able to
make authenticating with website2020 applications simpler for
app to app as it is just adding a header.

However this will require applications to now store the JWT
somewhere rather than fetching it from the cookie.